issue fixed InMemo implementation removed Tests refactored
