+++ .idea/libraries/Maven__org_postgresql_postgresql_9_4_1210.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="Maven: org.postgresql:postgresql:9.4.1210">
+    <CLASSES>
+      <root url="jar://$MAVEN_REPOSITORY$/org/postgresql/postgresql/9.4.1210/postgresql-9.4.1210.jar!/" />
+    </CLASSES>
+    <JAVADOC>
+      <root url="jar://$MAVEN_REPOSITORY$/org/postgresql/postgresql/9.4.1210/postgresql-9.4.1210-javadoc.jar!/" />
+    </JAVADOC>
+    <SOURCES>
+      <root url="jar://$MAVEN_REPOSITORY$/org/postgresql/postgresql/9.4.1210/postgresql-9.4.1210-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
