+++ src/main/java/server/model/data/UserProfile.java
@@ -84,4 +84,7 @@ public class UserProfile {
         this.name = name;
     }
 
+    public void setRegistrationDate(Date registrationDate) {
+        this.registrationDate = registrationDate;
+    }
 }
