--- src/test/java/server/model/dao/UserProfileInMemoTest.java
@@ -1,14 +1,0 @@
-package server.model.dao;
-
-import org.junit.Test;
-
-import static org.junit.Assert.*;
-
-/**
- * Created by eugene on 10/22/16.
- */
-public class UserProfileInMemoTest extends UserDAOTest {
-    static {
-        dao = new UserProfileInMemo();
-    }
-}
